+++ tests/test_rook.py
@@ -12,39 +12,75 @@ class TestRook(unittest.TestCase):
         self.board = [[None for _ in range(8)] for _ in range(8)]
         self.__rb__ = Rook("black")
         self.__rw__= Rook("white")
-       
+        self.board = Board(for_test=True)
 
    
     def test_movimientos_verticales_con_compañeros_en_el_camino(self): #BLANCAS
-        self.board[4][1] = self.__rw__
-        self.board[2][1] = Piece("white")  
-        self.board[6][1] = Piece("white")
+        self.board.set_piece(4,1,self.__rw__)
+        self.board.set_piece(2,1 ,Piece("white"))  
+        self.board.set_piece(6 ,1 ,Piece("white"))
         resultados= [(3, 1), (5, 1)]
         self.assertEqual(self.__rw__.posiciones_verticales(4, 1, self.board), resultados)
 
     def test_movimientos_verticales_con_compañeros_en_el_camino(self): #NEGRAS
-        self.board[4][3] = self.__rb__
-        self.board[2][3] = Piece("black")  
-        self.board[6][3] = Piece("black")
+        self.board.set_piece(4,3 ,self.__rb__)
+        self.board.set_piece(2,3 , Piece("black"))  
+        self.board.set_piece(6,3 , Piece("black"))
         resultados= [(3, 3), (5, 3)]
         self.assertEqual(self.__rb__.posiciones_verticales(4, 3, self.board), resultados)
 
     def test_movimientos_horizontales_con_compañeros_en_el_camino(self): #BLANCAS
-        self.board[4][3] = self.__rw__
-        self.board[4][1] = Piece("white")  
-        self.board[4][5] = Piece("white")
+        self.board.set_piece(4,3,self.__rw__)
+        self.board.set_piece(4,1 , Piece("white"))  
+        self.board.set_piece(4,5 , Piece("white"))
         resultados= [(4, 2), (4, 4)]
         self.assertEqual(self.__rw__.posiciones_horizontales(4, 3, self.board), resultados)
 
     def test_movimientos_horizontales(self):  #BLANCAS
-        self.board[4][3] = self.__rw__
+        self.board.set_piece(4,3 ,self.__rw__)
         movimientos = [(4,0),(4,1),(4,2),(4,4),(4,5),(4,6),(4,7)]
         self.assertEqual(self.__rw__.posiciones_horizontales(4, 3, self.board), movimientos)
 
     def test_movimientos_verticales(self):  #BLANCAS
-        self.board[4][3] = self.__rw__
+        self.board.set_piece(4,3 ,self.__rw__)
         movimientos = [(0,3),(1,3),(2,3),(3,3),(5,3),(6,3),(7,3)]
         self.assertEqual(self.__rw__.posiciones_verticales(4, 3, self.board), movimientos)
+
+    def test_movimientos_horizontales_con_captura(self): #BLANCAS
+        self.board.set_piece(2,3,self.__rw__)
+        self.board.set_piece(2,1 , Piece("black"))  
+        self.board.set_piece(2,5 , Piece("black"))
+        resultados= [(2,2),(2,1),(2,4),(2,5)]
+        resultados_obtenidos = sorted(self.__rw__.posiciones_horizontales(2, 3, self.board))
+        self.assertEqual(sorted(resultados_obtenidos), sorted(resultados))
+
+    def test_movimientos_verticales_con_captura(self): #BLANCAS
+        self.board.set_piece(2,3,self.__rw__)
+        self.board.set_piece(0,3 , Piece("black"))  
+        self.board.set_piece(4,3 , Piece("black"))
+        resultados= [(1,3),(0,3),(4,3),(3,3)]
+        resultados_obtenidos = sorted(self.__rw__.posiciones_verticales(2, 3, self.board))
+        self.assertEqual(sorted(resultados_obtenidos), sorted(resultados))
+
+    def test_movimientos_verticales_con_captura_y_compañero(self): #BLANCAS
+        self.board.set_piece(2,3,self.__rw__)
+        self.board.set_piece(0,3 , Piece("black"))  
+        self.board.set_piece(1,3 , Piece("white"))
+        self.board.set_piece(3,3 , Piece("black"))
+        resultados= [(3,3)]
+        resultados_obtenidos = sorted(self.__rw__.posiciones_verticales(2, 3, self.board))
+        self.assertEqual(sorted(resultados_obtenidos), sorted(resultados))
+
+    def test_movimientos_horizontal_con_captura_y_compañero(self): #BLANCAS
+        self.board.set_piece(2,3,self.__rw__)
+        self.board.set_piece(2,1 , Piece("black"))  
+        self.board.set_piece(2,2 , Piece("white"))
+        self.board.set_piece(2,4 , Piece("black"))
+        resultados= [(2,4)]
+        resultados_obtenidos = sorted(self.__rw__.posiciones_horizontales(2, 3, self.board))
+        self.assertEqual(sorted(resultados_obtenidos), sorted(resultados))
+    
+